requirements.txt
faker
psycopg2-binary
confluent_kafka
simplejson